requirements.txt
numpy
#torch>=1.1
#spconv
numba
tensorboardX
easydict
pyyaml
packaging
protobuf
tqdm
torchinfo
scikit-image
